requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.58.2
openpyxl==3.1.2
ruff==0.1.6
gunicorn==21.2.0
Pillow==10.1.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
uvicorn==0.24.0.post1